==> Análise de sentimento/1. Classificacao etapa 1.ipynb <==
import pandas as pd
import numpy as np
import os as os
import pickle
import seaborn as sns
import matplotlib.pyplot as plt
from wordcloud import WordCloud, STOPWORDS, ImageColorGenerator
from nltk.util import ngrams
import nltk
from arabica import arabica_freq
from unidecode import unidecode
import re
import plotly.express as px
import texthero as hero
from sklearn.cluster import KMeans
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import LatentDirichletAllocation
from sklearn.decomposition import NMF
from sklearn.decomposition import TruncatedSVD
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from sklearn.naive_bayes import BernoulliNB
import spacy
import re
import nlp
import warnings
import statsmodels.api as sm
from statsmodels.formula.api import ols


warnings.simplefilter(action='ignore', category=FutureWarning)

df = pd.read_csv('sentimento - base1.csv')

df.head()

df['ds_message'][236]

def classif_emoji(text):
    '''
    Replace emojis with word emoji
    '''
    regrex_pattern = re.compile(pattern="["
                                u"\U0001F600-\U0001F64F"  # emoticons
                                u"\U0001F300-\U0001F5FF"  # symbols & pictographs
                                u"\U0001F680-\U0001F6FF"  # transport & map symbols
                                u"\U0001F1E0-\U0001F1FF"  # flags (iOS)
                                "]+", flags=re.UNICODE)
    return regrex_pattern.sub(r'emoji ', str(text))

def clean_text(x):
    x = unidecode(x)
    x = x.strip()
    x = x.lower()
    x = re.sub(r'\t', ' ', x)
    x = re.sub(r'http\S+', '', x)  # remove url
    x = re.sub(r'screenshot sent', 'screenshotsent', x)
    x = re.sub(r'will bank', 'willbank', x)
    x = re.sub(r'(nbsp\;|[rl]dquo\;|quot\;|acute\;|tilde\;|circ\;|grave\;|cedil\;|\&|\/)', '', x)  # replace special characters
    
    x = x.replace('&ccedil;','ç')
    x = x.replace('&atilde;;','ã')
    x = x.replace('&otilde;;','õ')
    x = re.sub(r'(\d{1,}|\!|\?|\,|\.|\\n|\:|\;|\`|\´|\-|\_)', ' ', x)  # replace punctuation and remove numbers
    x = x.replace('  ',' ')
    x = x.replace('\n','')
    x = x.strip()
    return x


len(df)

# não vou remover emojis
df['ds_message_tratada'] = df['ds_message'].apply(clean_text)

df['ds_message_tratada'][132]

df['n_palavras'] = df['ds_message_tratada'].str.count(' ')

from LeIA import SentimentIntensityAnalyzer 


# Initialize the sentiment analyzer
analyzer = SentimentIntensityAnalyzer()
# Create a new column in the dataframe to store the sentiment scores
df["sentiment"] = 0.0

# Iterate over the rows of the dataframe
for index, row in df.iterrows():
  # Use the analyzer to predict the sentiment of the text
  scores = analyzer.polarity_scores(row["ds_message"])
  # Store the compound score in the new column
  df.at[index, "sentiment"] = scores["compound"]

df['sentiment']

df[['sentiment', 'SENTIMENTO']].groupby('SENTIMENTO').mean()

df['sentiment'].describe()

len(df.query('sentiment < -0.9'))

df.query('sentiment < -0.9').to_csv('muito_negativo_alg_leya.csv')

df['fx_sentiment'] = np.where(df.sentiment < 0, 'negativo', 
                     np.where(df.sentiment > 0, 'positivo', 'neutro'))

df['fx_sentiment'].value_counts()

pd.crosstab(df.fx_sentiment, df.SENTIMENTO)

for i in df.query('fx_sentiment == "negativo" & SENTIMENTO == "positivo"').sort_values('sentiment', ascending = True).head(10)['ds_message']:
    print(i)

for i in df.query('fx_sentiment == "negativo" & SENTIMENTO == "neutro"').sort_values('sentiment', ascending = True).head(10)['ds_message']:
    print(i)

for i in df.query('fx_sentiment == "positivo" & SENTIMENTO == "negativo"').sort_values('sentiment', ascending = True).head(10)['ds_message']:
    print(i)

for i in df.query('sentiment < 0')[['sentiment', 'SENTIMENTO', 'ds_message']].sort_values('sentiment', ascending = True).head(10)['ds_message']:
    print(i)

exemplo_base = df[df['SENTIMENTO'].notnull()]
len(exemplo_base)

stopwords_will= ['da', 'em', 'você', 'de', 'ao', 'os', 'que', 'e', 'eu', 'o', 'estou', 'ele','ela','pra', 'pro', 'entao', 'para', 'voce',
                 'pra', 'tá', 'né', 'ah', 'aí', 'dá', 'ó', 'mim', 'então', 'aqui', 'tô', 'pro',
            'seu', 'sua', 'por', 'sobre', 'mais', 'uma', 'um','como', 'meu', 'com', 'outro','outra','das','dos','foi', 'fiz','pelo','tem','mas',
            'este', 'esse','ta', 'to', 'ai', 'la', 'lo', 'lhe', 'ne','pra', 'tá', 'né', 'ah', 'aí', 'dá', 'ó', 'mim', 'então', 'aqui', 'tô', 'pro', 'isso']

exemplo_base['ds_message'] = exemplo_base['ds_message'].fillna("").apply(lambda x: ' '.join([word for word in x.split() if word not in stopwords_will]))
## retirando sinais de pontuação
exemplo_base['ds_message'] = exemplo_base['ds_message'].str.replace('[^\w\s]','')
## colocando todas as letras das mensagens em letras minúsculas
exemplo_base['ds_message'] = exemplo_base['ds_message'].str.lower() 

## Análise de sentimento
vetorizador = CountVectorizer()

X = vetorizador.fit_transform(exemplo_base['ds_message'])

#X = np.append(X, exemplo_base['n_palavras'])


y = exemplo_base.SENTIMENTO

clf = BernoulliNB()
clf = clf.fit(X,y)

def sentimentos(df, ds_message_tratada):
    df['sentimento_predict'] = clf.predict(vetorizador.transform(df[ds_message_tratada].fillna('')))
    return df

exemplo_base = sentimentos(exemplo_base, 'ds_message')

exemplo_base.head()

exemplo_base.sentimento_predict.value_counts()

pd.crosstab(exemplo_base.sentimento_predict, exemplo_base.SENTIMENTO)

for i in exemplo_base.query('sentimento_predict == "neutro" & SENTIMENTO == "negativo"').head(10)['ds_message']:
    print(i)

#Modelo de duvida, problema ou solicitacao

y = exemplo_base.TIPO_MSG

clf2 = BernoulliNB()
clf2 = clf2.fit(X,y)

def tipo_msg(df, ds_message_tratada):
    df['tipo_msg_predict'] = clf2.predict(vetorizador.transform(df[ds_message_tratada].fillna('')))
    return df

exemplo_base = tipo_msg(exemplo_base, 'ds_message')


exemplo_base.tipo_msg_predict.value_counts()

pd.crosstab(exemplo_base.tipo_msg_predict, exemplo_base.TIPO_MSG)

for i in exemplo_base.query('tipo_msg_predict == "duvida" & TIPO_MSG == "problema"').head(10)['ds_message']:
    print(i)

## Próximos passos

- Focar no preenchimento dos casos mapeados pelo Leya como muito negativos (checar)
- Procurar os casos com palavras de baixo calão
- Procurar casos "já faz", "problema" para acerto da classificação de problemas
- Avaliar se em alguns casos a classificação de sentimento negativo não seria neutro de fato (clientes que relatam demora ou que passou do prazo seria um fato objetivo ou escondem um stress?)
- Testar classificador usando um embedding ao invés de vetorizer
- Incluir nr de palavras e outras característias semânticas
