==> Análise de sentimento/4. Classificacao etapa 3 com Stacking.ipynb <==
import pandas as pd
import numpy as np
import os as os
import pickle
import seaborn as sns
import matplotlib.pyplot as plt
from wordcloud import WordCloud, STOPWORDS, ImageColorGenerator
from nltk.util import ngrams
import nltk
from arabica import arabica_freq
from unidecode import unidecode
import re
import plotly.express as px
import texthero as hero
from sklearn.cluster import KMeans
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import LatentDirichletAllocation
from sklearn.decomposition import NMF
from sklearn.decomposition import TruncatedSVD
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from sklearn.naive_bayes import BernoulliNB
import spacy
import re
import nlp
import warnings
import statsmodels.api as sm
from statsmodels.formula.api import ols


warnings.simplefilter(action='ignore', category=FutureWarning)

df = pd.read_csv('sentimento - base3.csv')

df.head()

# def classif_emoji(text):
#     '''
#     Replace emojis with word emoji
#     '''
#     regrex_pattern = re.compile(pattern="["
#                                 u"\U0001F600-\U0001F64F"  # emoticons
#                                 u"\U0001F300-\U0001F5FF"  # symbols & pictographs
#                                 u"\U0001F680-\U0001F6FF"  # transport & map symbols
#                                 u"\U0001F1E0-\U0001F1FF"  # flags (iOS)
#                                 "]+", flags=re.UNICODE)
#     return regrex_pattern.sub(r'emoji ', str(text))

def clean_text(x):
    #x = unidecode(x)
    x = x.strip()
    x = x.lower()
    x = re.sub(r'\t', ' ', x)
    x = re.sub(r'http\S+', '', x)  # remove url
    x = re.sub(r'screenshot sent', 'screenshotsent', x)
    x = re.sub(r'will bank', 'willbank', x)
    x = re.sub(r'(nbsp\;|[rl]dquo\;|quot\;|acute\;|tilde\;|circ\;|grave\;|cedil\;|\&|\/)', '', x)  # replace special characters
    x = x.replace('&ccedil;','ç')
    x = x.replace('&atilde;;','ã')
    x = x.replace('&otilde;;','õ')
    #x = re.sub(r'(\d{1,}|\!|\?|\,|\.|\\n|\:|\;|\`|\´|\-|\_)', ' ', x)  # replace punctuation and remove numbers
    x = x.replace('  ',' ')
    x = x.replace('\n','')
    x = x.strip()
    return x


len(df)

# não vou remover emojis
df['ds_message_tratada'] = df['ds_message'].apply(clean_text)

df['ds_message_tratada'][4]

df['n_palavras'] = df['ds_message_tratada'].str.count(' ')

df['n_palavras_t'] = np.where(df.n_palavras >= 20, 20, df.n_palavras)

pd.crosstab(df.n_palavras_t, df.SENTIMENTO)

from LeIA import SentimentIntensityAnalyzer 


# Initialize the sentiment analyzer
analyzer = SentimentIntensityAnalyzer()
# Create a new column in the dataframe to store the sentiment scores
df["sentiment"] = 0.0

# Iterate over the rows of the dataframe
for index, row in df.iterrows():
  # Use the analyzer to predict the sentiment of the text
  scores = analyzer.polarity_scores(row["ds_message"])
  # Store the compound score in the new column
  df.at[index, "sentiment"] = scores["compound"]

exemplo_base = df[df['SENTIMENTO'].notnull()]
len(exemplo_base)

stopwords_will= ['da', 'em', 'você', 'de', 'ao', 'os', 'que', 'e', 'eu', 'o', 'estou', 'ele','ela','pra', 'pro', 'entao', 'para', 'voce',
                 'pra', 'tá', 'né', 'ah', 'aí', 'dá', 'ó', 'mim', 'então', 'aqui', 'tô', 'pro',
            'seu', 'sua', 'por', 'sobre', 'mais', 'uma', 'um','como', 'meu', 'com', 'outro','outra','das','dos','foi', 'fiz','pelo','tem','mas',
            'este', 'esse','ta', 'to', 'ai', 'la', 'lo', 'lhe', 'ne','pra', 'tá', 'né', 'ah', 'aí', 'dá', 'ó', 'mim', 'então', 'aqui', 'tô', 'pro', 'isso']

exemplo_base['SENTIMENTO_label'] = np.where(exemplo_base['SENTIMENTO'] == "negativo", 0, 
                                            np.where(exemplo_base['SENTIMENTO'] == "neutro", 1, 2 ))

exemplo_base['SENTIMENTO_label'].value_counts()

from sklearn.model_selection import train_test_split

vetorizador = TfidfVectorizer()
X = vetorizador.fit_transform(exemplo_base['ds_message'])
y = exemplo_base.SENTIMENTO_label


x_train, x_test, y_train, y_test = train_test_split(X
                                                    , exemplo_base.SENTIMENTO_label
                                                    , random_state=42, test_size=0.20)

import xgboost as xgb
from sklearn.model_selection import cross_val_score, StratifiedKFold, KFold
from sklearn import metrics

#INCLUIR PARAMETROS DO MELHOR MODELO
clf = xgb.XGBClassifier(objective="multi:softprob", max_depth = 3, num_round = 437, eta = 0.02, subsample = 0.8,  alpha = 0.0, colsample_bytree = 0.40, gamma = 3.7, random_state = 42)
#lambda = 1.644,
clf = clf.fit(x_train,y_train)

y_predicted = clf.predict(x_test)
pd.crosstab(y_predicted, y_test)

scores = cross_val_score(clf, x_train, y_train, cv=5, scoring='f1_macro')
print("%0.2f F1 with a standard deviation of %0.2f" % (scores.mean(), scores.std()))

from sklearn.metrics import classification_report

print(classification_report(y_test, y_predicted))


import xgboost as xgb
from sklearn.model_selection import cross_val_score, StratifiedKFold, KFold
from sklearn import metrics

#INCLUIR PARAMETROS DO MELHOR MODELO
clf = xgb.XGBClassifier(objective="multi:softprob", max_depth = 10)
#lambda = 1.644,
clf = clf.fit(x_train,y_train)

y_predicted = clf.predict(x_test)
pd.crosstab(y_predicted, y_test)

print(classification_report(y_test, y_predicted))

#INCLUIR PARAMETROS DO MELHOR MODELO
clf = xgb.XGBClassifier(objective="multi:softprob", max_depth = 30)
#lambda = 1.644,
clf = clf.fit(x_train,y_train)

y_predicted = clf.predict(x_test)
pd.crosstab(y_predicted, y_test)

print(classification_report(y_test, y_predicted))

#INCLUIR PARAMETROS DO MELHOR MODELO
clf = xgb.XGBClassifier(objective="multi:softprob", max_depth = 50)
#lambda = 1.644,
clf = clf.fit(x_train,y_train)

y_predicted = clf.predict(x_test)
pd.crosstab(y_predicted, y_test)

print(classification_report(y_test, y_predicted))

from sklearn.neural_network import MLPClassifier

vetorizador = TfidfVectorizer()
X = vetorizador.fit_transform(exemplo_base['ds_message_tratada'])
y = exemplo_base.SENTIMENTO_label

x_train, x_test, y_train, y_test = train_test_split(X
                                                    , exemplo_base.SENTIMENTO_label
                                                    , random_state=42, test_size=0.20)

clf = MLPClassifier(solver='lbfgs', alpha=0.0001, hidden_layer_sizes = 100, random_state=1) #alpha=0.0001

clf = clf.fit(x_train, y_train)
y_predicted = clf.predict(x_test)
pd.crosstab(y_predicted, y_test)

print(classification_report(y_test, y_predicted))

vetorizador = TfidfVectorizer()
X = vetorizador.fit_transform(exemplo_base['ds_message_tratada'])
y = exemplo_base.SENTIMENTO_label

x_train, x_test, y_train, y_test = train_test_split(X
                                                    , exemplo_base.SENTIMENTO_label
                                                    , random_state=42, test_size=0.20)

clf = MLPClassifier(solver='lbfgs', alpha=0.0001, hidden_layer_sizes = 200, random_state=1) #alpha=0.0001

clf = clf.fit(x_train, y_train)
y_predicted = clf.predict(x_test)
pd.crosstab(y_predicted, y_test)

print(classification_report(y_test, y_predicted))

vetorizador = TfidfVectorizer()
X = vetorizador.fit_transform(exemplo_base['ds_message_tratada'])
y = exemplo_base.SENTIMENTO_label

x_train, x_test, y_train, y_test = train_test_split(X
                                                    , exemplo_base.SENTIMENTO_label
                                                    , random_state=42, test_size=0.20)

clf = MLPClassifier(solver='lbfgs', alpha=0.0001, hidden_layer_sizes = 50, random_state=1) #alpha=0.0001

clf = clf.fit(x_train, y_train)
y_predicted = clf.predict(x_test)
pd.crosstab(y_predicted, y_test)

print(classification_report(y_test, y_predicted))

- Combinar MLP com 100 hl + XGBoost com max_depth = 30 + score Leya

## Início Stacking

mlp = MLPClassifier(solver='lbfgs', alpha=0.0001, hidden_layer_sizes = 100, random_state=1) #alpha=0.0001
mlp = mlp.fit(x_train, y_train)
tree = xgb.XGBClassifier(objective="multi:softprob", max_depth = 30)
tree = tree.fit(x_train, y_train)


x_test2_pr_mlp_0 = mlp.predict_proba(x_test)[:, 0]
x_test2_pr_mlp_1 = mlp.predict_proba(x_test)[:, 1]
x_test2_pr_mlp_2 = mlp.predict_proba(x_test)[:, 2]

x_test2_pr_tree_0 = tree.predict_proba(x_test)[:, 0]
x_test2_pr_tree_1 = tree.predict_proba(x_test)[:, 1]
x_test2_pr_tree_2 = tree.predict_proba(x_test)[:, 2]


x_test2 = pd.DataFrame()

x_test2['prob_mlp_0'] = x_test2_pr_mlp_0
x_test2['prob_mlp_1'] = x_test2_pr_mlp_1
x_test2['prob_mlp_2'] = x_test2_pr_mlp_2

x_test2['prob_tree_0'] = x_test2_pr_tree_0
x_test2['prob_tree_1'] = x_test2_pr_tree_1
x_test2['prob_tree_2'] = x_test2_pr_tree_2

x_test2 = x_test2.join(exemplo_base[['n_palavras_t', 'sentiment']])

x_test2['n_palavras_t'] = x_test2.n_palavras_t.fillna(0.0)
x_test2['sentiment'] = x_test2.sentiment.fillna(0.0)

x_test2.tail(5)

x_train2_pr_mlp_0 = mlp.predict_proba(x_train)[:, 0]
x_train2_pr_mlp_1 = mlp.predict_proba(x_train)[:, 1]
x_train2_pr_mlp_2 = mlp.predict_proba(x_train)[:, 2]

x_train2_pr_tree_0 = tree.predict_proba(x_train)[:, 0]
x_train2_pr_tree_1 = tree.predict_proba(x_train)[:, 1]
x_train2_pr_tree_2 = tree.predict_proba(x_train)[:, 2]

x_train2 = pd.DataFrame()

x_train2['prob_mlp_0'] = x_train2_pr_mlp_0
x_train2['prob_mlp_1'] = x_train2_pr_mlp_1
x_train2['prob_mlp_2'] = x_train2_pr_mlp_2

x_train2['prob_tree_0'] = x_train2_pr_tree_0
x_train2['prob_tree_1'] = x_train2_pr_tree_1
x_train2['prob_tree_2'] = x_train2_pr_tree_2

x_train2 = x_train2.join(exemplo_base[['n_palavras_t', 'sentiment']])

x_train2['n_palavras_t'] = x_train2.n_palavras_t.fillna(0.0)
x_train2['sentiment'] = x_train2.sentiment.fillna(0.0)

x_train2.head(5)

clf = xgb.XGBClassifier(objective="multi:softprob", max_depth = 10, random_state = 42)
clf = clf.fit(x_train2, y_train)

y_predicted = []
y_predicted = clf.predict(x_test2)
pd.crosstab(y_predicted, y_test)

print(classification_report(y_test, y_predicted))
