==> Análise de sentimento/6. Regras fixas para Problema-Duvida-Reclamacao.ipynb <==
from matplotlib import pyplot as plt
from apyori import apriori
import pandas as pd
import numpy as np
import aed as aed

df = pd.read_csv('De para tag.csv')

df2 = pd.concat([df]*100).reset_index()

len(df2)

df2.head()

df2['n_geral'] = df2['n1'] + '|' + df2['n2'] + '|' + df2['n3']

### N1 POR TIPO DE TAG

df_assoc = df2[['n1', 'tipo_de_tag']]

transactions = []
for i in range(0, df_assoc.shape[0]):
    transactions.append([str(df_assoc.values[i, j]) for j in range(0, 2)])

print(transactions[0])

rules = apriori(transactions, min_support = 0.01, min_confidence = 0.10, min_lift = 1, min_length = 2)
results = list(rules)
results = pd.DataFrame(results)
results.head(30)

###  relacionando com n2

feat = ['n2', 'tipo_de_tag']

df_assoc = df2[feat]

transactions = []
for i in range(0, df_assoc.shape[0]):
    transactions.append([str(df_assoc.values[i, j]) for j in range(0, 2)])

print(transactions[0])

rules = apriori(transactions, min_support = 0.01, min_confidence = 0.10, min_lift = 1, min_length = 2)
results = list(rules)
results = pd.DataFrame(results)
results.head(50)

### relacionando com n3

feat = ['n3', 'tipo_de_tag']

df_assoc = df2[feat]

transactions = []
for i in range(0, df_assoc.shape[0]):
    transactions.append([str(df_assoc.values[i, j]) for j in range(0, 2)])

print(transactions[0])

rules = apriori(transactions, min_support = 0.01, min_confidence = 0.10, min_lift = 1, min_length = 2)
results = list(rules)
results = pd.DataFrame(results)
results.head(50)

feat = ['n1', 'tipo_de_tag']
drop_col = []
aed.find_correl(df[feat], drop_col, 0.30)

feat = ['n2', 'tipo_de_tag']
drop_col = []
aed.find_correl(df[feat], drop_col, 0.30)

feat = ['hierarquia_1', 'tipo_de_tag']
drop_col = []
aed.find_correl(df[feat], drop_col, 0.30)

feat = ['hierarquia_2', 'tipo_de_tag']
drop_col = []
aed.find_correl(df[feat], drop_col, 0.30)

### Função de combinações

import funcionalidade as fc

df['cont'] = 1

colunas_analise = ['n1', 'n2','n3','hierarquia_1','hierarquia_2','hierarquia_3','hierarquia_4', 'flag_fraude']

report_combinacoes = fc.report_perfil_grupos(df, 'tipo_de_tag', 'cont', 0.30, 10, colunas_analise)
report_combinacoes.set_index('tipo_de_tag', inplace=True)

report_combinacoes.to_csv('regras_hierarquia.csv')

colunas_analise = ['hierarquia_1','hierarquia_2','hierarquia_3','hierarquia_4', 'flag_fraude']

report_combinacoes = fc.report_perfil_grupos(df, 'tipo_de_tag', 'cont', 0.30, 10, colunas_analise)
report_combinacoes.set_index('tipo_de_tag', inplace=True)

report_combinacoes.to_csv('regras_hierarquia_2.csv')

### NLP

import re
import string 
import unidecode
from numpy.core.defchararray import find
from numpy.core.defchararray import replace

def text_preproc(x):
    if x is not None:
        x = x.upper()
        x = unidecode.unidecode(x)
        x = x.replace(","," ")
        x = re.sub(r'\s', ' ', x)
        x = re.sub(r'\r', ' ', x)
        x = re.sub(r'\t', ' ', x)
        x = re.sub(r"'", ' ', x)
        x = re.sub(r'/[^a-zA-Z0-9 ]/g', ' ', x)
        x = re.sub(r'/[^\w\s]/gi', '', x)
        x = x.replace('     ',' ')
        x = x.replace('   ',' ')
        x = x.replace('  ',' ')
        x = x.strip()

        return x
    
    else:
        return None
    
def count_words_cluster(df, termo, coluna_cluster):
    contagemClusters = pd.DataFrame(columns=['index', 'count', 'cluster'])
    clusters = df[coluna_cluster].unique()
    for i in clusters:
        filtro_cluster = df[df[coluna_cluster] == i]
        newdf = filtro_cluster[termo].str.split(expand=True).stack().value_counts().reset_index(name="count")
        newdf['cluster'] = i
        contagemClusters = contagemClusters.append(newdf)
    return contagemClusters

df['N3_T'] = df['n3'].fillna('').apply(text_preproc)

df_cont = count_words_cluster(df, 'N3_T', 'tipo_de_tag' )
df_cont = df_cont.query('count > 1').sort_values(['cluster', 'count'], ascending=[True, False]).groupby('index').agg({'cluster':['first'],'count':['first'] })
df_cont = df_cont.reset_index()
df_cont = df_cont.sort_values(('count', 'first'), ascending = False)

retirar = ["da", "de", "o", "a", "outros", "em", "do", "na", "nas", "/", "e", "E", "DE", "DA", "NAO", "PARA"]
df_cont_final = df_cont[~df_cont['index'].isin(retirar)]
df_cont_final.head(20)

df_cont_final.to_csv('palavras_chave.csv')

## Arvore de decisão

from sklearn.tree import DecisionTreeClassifier 
from sklearn import tree

df_dm = pd.get_dummies(df2[['hierarquia_2', 'tipo_de_tag']], columns = ['hierarquia_2', 'tipo_de_tag'])

df_dm.head()

for i in df_dm.columns:   
    print(i)
    X = np.array(df_dm[df_dm.drop(i, axis = 1).columns]).reshape(1, -1)
    y = np.array(df_dm[i]).reshape(1, -1)
    X.shape
    
    clf = DecisionTreeClassifier(max_depth=2, random_state=1234)
    model = clf.fit(X, y)

    text_representation = tree.export_text(clf)

    print(text_representation)
    print("_________________")
