==> Análise de sentimento/7. Salva joblib-MLP.ipynb <==
import pandas as pd
import numpy as np
import os as os
#import pickle
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn import metrics
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score, StratifiedKFold, KFold, RepeatedKFold
from sklearn.metrics import classification_report
from joblib import dump, load

import warnings

warnings.simplefilter(action='ignore', category=FutureWarning)

df = pd.read_csv('sentimento - base3.csv')

df.head()

len(df)

exemplo_base = df[df['SENTIMENTO'].notnull()]
len(exemplo_base)

stopwords_will= ['da', 'em', 'você', 'de', 'ao', 'os', 'que', 'e', 'eu', 'o', 'estou', 'ele','ela','pra', 'pro', 'entao', 'para', 'voce',
                 'pra', 'tá', 'né', 'ah', 'aí', 'dá', 'ó', 'mim', 'então', 'aqui', 'tô', 'pro',
            'seu', 'sua', 'por', 'sobre', 'mais', 'uma', 'um','como', 'meu', 'com', 'outro','outra','das','dos','foi', 'fiz','pelo','tem','mas',
            'este', 'esse','ta', 'to', 'ai', 'la', 'lo', 'lhe', 'ne','pra', 'tá', 'né', 'ah', 'aí', 'dá', 'ó', 'mim', 'então', 'aqui', 'tô', 'pro', 'isso']

### Modelo Sentimento

exemplo_base['SENTIMENTO_label'] = np.where(exemplo_base['SENTIMENTO'] == "negativo", 0, 
                                            np.where(exemplo_base['SENTIMENTO'] == "neutro", 1, 2 ))

exemplo_base['SENTIMENTO_label'].value_counts()

from sklearn.neural_network import MLPClassifier

vetorizador = TfidfVectorizer()
X = vetorizador.fit_transform(exemplo_base['ds_message'])
y = exemplo_base.SENTIMENTO_label


x_train, x_test, y_train, y_test = train_test_split(X
                                                    , exemplo_base.SENTIMENTO_label
                                                    , random_state=42, test_size=0.20)

# Study OPTUNA: {'hidden_layer_sizes': 130, 'activation': 'relu', 'alpha': 0.18641747634500205}
clf_sentiment = MLPClassifier(solver='lbfgs', alpha=0.18641747634500205, hidden_layer_sizes= 130, activation = 'relu', warm_start = True)
clf_sentiment.fit(x_train, y_train)

y_predicted = clf_sentiment.predict(x_test)
pd.crosstab(y_predicted, y_test)

print(classification_report(y_test, y_predicted))

dump(clf_sentiment, 'modelo_sentimento.joblib') 

dump(vetorizador, 'vetorizador_sentimento.joblib') 

# No Sagemaker
#clf = load('filename.joblib') 

### Avaliando o acerto com as palavras chave

base_predita = exemplo_base[['ds_message', 'SENTIMENTO_label']].copy()
base_predita['sent_predito_1'] = clf_sentiment.predict(X)

base_predita.head()

### Palavras chave 

chave_negativo = ['falta de respeito', 
'judicial',
'procon',
'jurídico',
'porra',
'merda', 
'palhaçada',
'carai',
'caralho',
'bosta',
'vocês demoram',
'vcs demoram',
'horrivel',
'banco ruim',
'péssimo',
'danos morais',
'pelo amor de deus',
'eu preciso resolver',
'me roubando']


chave_positivo = [
'ótimo',
'perfeito',
'muito obrigada',
'excelente',
'ótimo dia',
'maravilha',
'deus abençoe']

base_predita['sent_predito_1'].value_counts()

def marca_negativo(x):
    for i in chave_negativo:
        if (x.find(i) != -1):
            return 1
        else:
            return 0
   
base_predita['flag_negativo'] = base_predita['ds_message'].apply(marca_negativo)

def marca_positivo(x):
    for i in chave_positivo:
        if (x.find(i) != -1):
            return 1
        else:
            return 0
   
base_predita['flag_positivo'] = base_predita['ds_message'].apply(marca_positivo)

base_predita['sent_predito_2'] = np.where(base_predita['flag_negativo'] == 1,  0, #condições negativo
                                 np.where(base_predita['flag_positivo'] == 1,  2, #condições positivo
                                          base_predita['sent_predito_1']))

base_predita['sent_predito_2'].value_counts()

print(classification_report(base_predita['SENTIMENTO_label'], base_predita['sent_predito_1']))

print(classification_report(base_predita['SENTIMENTO_label'], base_predita['sent_predito_2']))

### Modelo do tipo de mensagem

exemplo_base['TIPO_MSG'].value_counts()

exemplo_base['TIPO_MSG_label'] = np.where(exemplo_base['TIPO_MSG'] == "problema", 0, 
                                            np.where(exemplo_base['TIPO_MSG'] == "solicitacao", 1, 2 ))

exemplo_base['TIPO_MSG_label'].value_counts()


vetorizador = TfidfVectorizer()
X = vetorizador.fit_transform(exemplo_base['ds_message'])
y = exemplo_base.TIPO_MSG_label

x_train, x_test, y_train, y_test = train_test_split(X
                                                    , exemplo_base.TIPO_MSG_label
                                                    , random_state=42, test_size=0.20)

from sklearn.neural_network import MLPClassifier
clf_tipo_msg = MLPClassifier(solver='lbfgs', alpha=3.4372969952170234, hidden_layer_sizes= 160, activation = 'tanh', warm_start = True)
clf_tipo_msg.fit(x_train, y_train)

y_predicted = clf_tipo_msg.predict(x_test)
pd.crosstab(y_predicted, y_test)

print(classification_report(y_test, y_predicted))

dump(clf_tipo_msg, 'modelo_tipo_mensagem.joblib')
