==> Análise de sentimento/9. Aplica joblibs pesquisa CSAT.ipynb <==
import pandas as pd
import numpy as np
import os as os
#import pickle
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn import metrics
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score, StratifiedKFold, KFold, RepeatedKFold
from sklearn.metrics import classification_report
from joblib import dump, load
import acessos as ac
import warnings

warnings.simplefilter(action='ignore', category=FutureWarning)

QUERY 

DIA MÍNIMO E MÁXIMO DE INGESTÃO 
- 2023-09-29 16:47:45.401
- 2023-11-24 02:20:19.067


query = """select 
id_conversation
, ingestion_date
, ds_comentario
from processed_zone_events_intercom.issues_ratings_intercom
where ds_comentario is not null"""

df = ac.df_athena_q('flavia-costa', query)

df.head()

len(df)

## Aplicando modelo sentimento

from sklearn.neural_network import MLPClassifier


vetorizador = load('vetorizador_sentimento.joblib') 

X = vetorizador.transform(df['ds_comentario'])

clf_sentiment = load('modelo_sentimento.joblib') 

df['sent_predito_1'] = clf_sentiment.predict(X)

### Palavras chave 

chave_negativo = ['falta de respeito', 
'judicial',
'procon',
'jurídico',
'porra',
'merda', 
'palhaçada',
'carai',
'caralho',
'bosta',
'vocês demoram',
'voces demoram',
'vcs demoram',
'horrivel',
'banco ruim',
'péssimo',
'danos morais',
'pelo amor de deus',
'eu preciso resolver',
'atender??',
'me roubando']


não fui respondid
não tive resposta
não resolveu o meu problema
meu problema ainda não foi resolvido
não ajudou

chave_positivo = [
'ótimo',
'otimo',
'perfeito',
'muito obrigada',
'excelente',
'ótimo dia',
'maravilha',
'deus abençoe']

Acrescentar:

Atendimento maravilhoso
atendimento nota 10
ótimo atendimento
atendimento excelente

df['sent_predito_1'].value_counts()

def marca_negativo(x):
    for i in chave_negativo:
        if (x.find(i) != -1):
            return 1
        else:
            return 0
   
df['flag_negativo'] = df['ds_comentario'].apply(marca_negativo)

def marca_positivo(x):
    for i in chave_positivo:
        if (x.find(i) != -1):
            return 1
        else:
            return 0
   
df['flag_positivo'] = df['ds_comentario'].apply(marca_positivo)

Para a pesquisa do CSAT temos uma tendencia maior para comentários positivos portanto podemos alterar a regra, que está diferente para análise de sentimento do chat

df['sent_predito_final'] = np.where((df['flag_negativo'] == 1),  0, #condições negativo
                                 np.where(df['flag_positivo'] == 1,  2, #condições positivo
                                          df['sent_predito_1']))

df['sent_predito_final'].value_counts()

### Modelo do tipo de mensagem

Gostaria de saber
Como faço
como faço
problema
não reconheço
não fiz essa compra
não reconhecida
fui roubad
fui assaltad
limite não voltou
não consegui pagar
pix não chegou
compra que não fiz
nunca comprei
quero alterar meu
quero resolver
queria que fosse resolvido
preciso resolver
quero aumentar o meu limite
queria o aumento do meu cartao
fiquei na dúvida
paguei mas não aparece
prefiro parcelar

clf_tipo_msg = load('modelo_tipo_mensagem.joblib')

df['tipo_msg_predito'] = clf_tipo_msg.predict(X)

df['tipo_msg_predito'].value_counts()

df['sentimento_predito'] = np.where(df['sent_predito_final'] == 0,  '0. Negativo', 
                           np.where(df['sent_predito_final'] == 2,  '2. Positivo', '1. Neutro'))

df.query('sentimento_predito == "0. Negativo"')['ds_message'][27]

df['tipo_msg_predito'] = np.where(df['tipo_msg_predito'] == 0,  '0. Problema', 
                           np.where(df['tipo_msg_predito'] == 1,  '1. Solicitacao', '2. Duvida'))

### Avaliação final e exportação

df.head()

df.to_csv('sentimento_csat.csv', index = False)

query2 = """select 
id_conversation, 
ds_nota
from processed_zone_events_intercom.issues_ratings_intercom
where ds_comentario is not null"""

df_nota = ac.df_athena_q('flavia-costa', query2)

df_nota.head()

df = df.join(df_nota.set_index('id_conversation'), on = 'id_conversation', how = 'left')

df.head()

df.query('sent_predito_1 == 0').head()

pd.crosstab(df.ds_nota, df.sent_predito_final, normalize = 'index')

""" area = 'cx'
name = 'analise_sentimento_teste.csv'
filename = f'{area}.{name}'

df[['cpf_customer', 'mes_chat', 'dt_envio_mensagem',
       'ds_message', 'sent_predito_1','tipo_msg_predito', 'sentimento_predito']].to_csv(filename, index=False)

os.system(f'aws s3 cp {filename} s3://data-sandbox-zone-will-prod/upload/') """
